Began directory structure. Script almost complete, see note
Need the composition of asm commands MOVZ and MOVK, which are IM format instructions? I did not search too hard but it is not extremely simple to find good reference for the formats. I found one which gave a different space size, the given output text gives a movz R#, lsl # instruction and I don't know where all the numbers come from.